Initial commit of the node-crfpp repo
Commits the initial structure of the project. The entry point of this project
is node-crfpp.js which contains the exposed functionality of the underlying
native CRF++ library. When running CRF++ normally, the user would interact with
the crf_learn.exe and crf_test.exe program. The crf_learn program is used to
create a trained model from training and template files. The crf_test program
uses the trained model to parse and tag an input of tokens. Currently, only the
crf_test portion of the CRF++ library is exposed in this project. This contains
functions that provide the same output as crf_test, i.e. toString(), as well
as helper functions that return tagged output in a more structured format.

TODO:
Finish README
Finalize tests
Finalize package bundle