Fix issues with empty input
Fixes issue with the tagger being read after the vlevel was reset instead of before.

Also, updates the crfpp submodule to a forked repo containing a fix for a crash
that could occur when next() was called on the tagger before parse(). As implemented
in CrfTest, if empty or invalid input is given, parse() will not be called on the
tagger, but next() will be.